+++ crates/nu-cli/src/commands/classified/external.rs
@@ -561,8 +561,8 @@ mod tests {
         let cmd = ExternalBuilder::for_name("i_dont_exist.exe").build();
 
         let input = InputStream::empty();
-        let mut ctx = crate::cli::basic_evaluation_context()
-            .expect("There was a problem creating a basic context.");
+        let mut ctx =
+            basic_evaluation_context().expect("There was a problem creating a basic context.");
 
         assert!(
             run_external_command(cmd, &mut ctx, input, ExternalRedirection::Stdout)
