+++ crates/nu-cli/src/commands/default_context.rs
@@ -1,9 +1,10 @@
 use crate::prelude::*;
+use nu_engine::basic_evaluation_context;
 use nu_engine::whole_stream_command;
 use std::error::Error;
 
 pub fn create_default_context(interactive: bool) -> Result<EvaluationContext, Box<dyn Error>> {
-    let context = crate::cli::basic_evaluation_context()?;
+    let context = basic_evaluation_context()?;
 
     {
         use crate::commands::*;
