+++ crates/nu-cli/src/env.rs
@@ -1,4 +1,3 @@
-pub(crate) mod basic_host;
 pub(crate) mod directory_specific_environment;
 pub(crate) mod environment;
 pub(crate) mod environment_syncer;
