+++ crates/nu-engine/Cargo.toml
@@ -39,6 +39,7 @@ serde = {version = "1.0.115", features = ["derive"]}
 serde_json = "1.0.57"
 tempfile = "3.1.0"
 termcolor = "1.1.0"
+term_size = "0.3.2"
 
 [target.'cfg(unix)'.dependencies]
 umask = "1.0.0"
