+++ crates/nu-engine/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod basic_evaluation_context;
 pub mod basic_shell_manager;
 mod call_info;
 mod command_args;
@@ -14,6 +15,7 @@ pub mod plugin;
 pub mod shell;
 mod whole_stream_command;
 
+pub use crate::basic_evaluation_context::basic_evaluation_context;
 pub use crate::basic_shell_manager::basic_shell_manager;
 pub use crate::call_info::UnevaluatedCallInfo;
 pub use crate::command_args::{
